Clippy

+++ crates/l2/sequencer/utils.rs
@@ -3,13 +3,10 @@ use std::time::Duration;
 use tokio::time::sleep;
 
 pub async fn sleep_random(sleep_amount: u64) {
-    let random_noise: i64 = {
+    let random_noise: u64 = {
         let mut rng = rand::thread_rng();
         rng.gen_range(0..400) - 200
     };
 
-    sleep(Duration::from_millis(
-        (sleep_amount as i64 + random_noise) as u64,
-    ))
-    .await;
+    sleep(Duration::from_millis(sleep_amount + random_noise)).await;
 }
